requirements.txt
# Master requirements for X-Ray DAQ System
# Core Flask dependencies
flask>=2.3.0
gunicorn>=21.0.0
werkzeug>=2.3.0

# Redis dependencies
redis>=5.0.0

# Database and data processing
influxdb-client>=1.36.0
python-dateutil>=2.8.0

# Networking and async
aiohttp>=3.8.0
requests>=2.31.0

# Configuration and utilities
python-dotenv>=1.0.0

# UI framework
nicegui>=1.4.0